requirements.txt
requests
loguru
python-dotenv